Adding original license to docker2singularity.sh
When the docker2singularity.sh script was removed from its original version and added to this repo, the original LICENSE file included https://github.com/vsoch/singularity-tools/blob/master/LICENSE was not honored, and the git history erased to unlink it. This should be added back to give due credit - note that the original script was written in June 2016 but this repository created later that year. For record, this repo was created and announced on [August 16 2016](https://groups.google.com/a/lbl.gov/forum/#!msg/singularity/kBTuLHlxQv0/yr-VAh1lEQAJ) whereas the creation of the script and sharing was [much earlier.](https://groups.google.com/a/lbl.gov/d/msg/singularity/FA0KTSYHqv4/8_D0ElILAgAJ).